moving buildCliConfig and directoriesToMount to cmd/options
